feat: Install 'flatted' package to parse and stringify objects that have circular structure when using local storage.
The "Project" object has a circular structure because it contains
the "toDoItems" property which is an array containing "ToDo" objects
which in turn also contains a property "project" which is a reference
to the Project object creating a cycle.

The JSON's parse and stringify methods could not handle this type of
structure. Hence, the package "flatted" will be used instead.